recipe-finder/requirements.txt
Flask
cs50
requests
openai
werkzeug.security
